Fixes EF database issue with casting decimal to double (schema errors). Adds support to inventory page to detect resistor/capacitor values when adding. Added better package type UI. Added quick toggle for enabling/disabling auto part info search. Fixed CSS issues.

+++ Binner/Library/Binner.Model/CommonPart.cs
@@ -58,12 +58,12 @@
         /// <summary>
         /// Cost for part from this supplier
         /// </summary>
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
 
         /// <summary>
         /// Total Cost for part (order line item total cost)
         /// </summary>
-        public double TotalCost { get; set; }
+        public decimal TotalCost { get; set; }
 
         /// <summary>
         /// Cost currency
